+++ CHANGE_LOG.md
@@ -9,6 +9,14 @@ If this is a problem, use exact build numbers to track changes. As each build lo
 
 ### Development Changes
 
+## 3.#.4 - 6/1/2018
+### Runtime Changes
+Added: Slot validation for centrifuge
+Added: Slot validation for extractor
+Fixed: Slot validation for boiler
+
+### Development Changes
+
 ## 3.0.3 - 6/1/2018
 ### Runtime Changes
 * Added: Wrench color orange